float inside return

+++ python_variable_annotations/8-make_multiplier.py
@@ -13,6 +13,6 @@ def make_multiplier(multiplier: float) -> typing.Callable[[float], float]:
         '''
         Function doc
         '''
-        return n * multiplier
+        return float(n * multiplier)
     
     return multi
